Refactored uploadManuscript to be testable. Also allowed article.test.js to run in record
