+++ server/xpub-server/entities/manuscript/resolvers/uploadManuscript.test.js
@@ -12,6 +12,7 @@ const { createTables } = require('@pubsweet/db-manager')
 const mailer = require('@pubsweet/component-send-email')
 const startS3rver = require('../../../test/mock-s3-server')
 const { User, File, Manuscript } = require('@elifesciences/xpub-model')
+const scienceBeamApi = require('./scienceBeamApi')
 const { Mutation } = require('.')
 const { userData, badUserData } = require('./index.test.data')
 
@@ -122,11 +123,21 @@ describe('Manuscripts', () => {
         ),
         mimetype: 'application/pdf',
       }
+
+      jest.spyOn(scienceBeamApi, 'extractSemantics').mockRejectedValueOnce({
+        error: {
+          code: 'ETIMEDOUT',
+          connect: false,
+          message: 'timed out',
+        },
+      })
+
       const manuscript = await Mutation.uploadManuscript(
         {},
         { id, file, fileSize: 73947 },
         { user: profileId },
       )
+      expect(manuscript.meta.title).toBe('')
       expect(manuscript).toMatchObject({
         id,
         meta: { title: '' },
